feat(#73): GoogleMap 컴포넌트 분리
